requirements.txt
# Israel Public Transit - Real-Time Monitoring Platform
# Python Dependencies

# ── Kafka ─────────────────────────────────────────────────
kafka-python==2.0.2

# ── GTFS-RT Protocol Buffers ──────────────────────────────
gtfs-realtime-bindings==2.0.0    # Google GTFS-RT protobuf bindings
protobuf>=4.24.0

# ── HTTP ──────────────────────────────────────────────────
requests==2.31.0

# ── Data Processing ───────────────────────────────────────
pandas==2.1.0
pyarrow==13.0.0
geopy==2.4.1

# ── AWS / Storage ─────────────────────────────────────────
boto3==1.28.0
psycopg2-binary==2.9.7

# NOTE: Apache Airflow runs inside Docker — do not install locally.
# See docker-compose.yml for the Airflow service configuration.

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.0elasticsearch
